web/modal/requirements.txt
opencv-python-headless>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
modal>=0.55.0
